Refactor suggestion order datatype & Add order check to avoid wasting space
If the calculated order wastes more than PAGE_SIZE, then drop down to
next smaller order. Worst case drop down to order 0. This avoids using
an mTHP that in theory would fit the requested size, but due to alignment
would only partly cover the requested range, and therefore waste most of
it's space. This way, less than PAGE_SIZE is wasted in downward alignment.
Refactor order datatype to be a pointer and add it to vm_fault struct, so
that it can be passed all the way down to alloc_anon_folio.